feat: AlertDialogComponent. integrated with swap confirmation
