+++ src/components/base/Dialogs/AlertDialog.tsx
@@ -1,6 +1,5 @@
-import { useRef, useCallback, useEffect, useState } from "react";
+import { useEffect, useRef } from "react";
 import {
-  useDisclosure,
   Button,
   AlertDialog,
   AlertDialogOverlay,
@@ -8,101 +7,79 @@ import {
   AlertDialogHeader,
   AlertDialogBody,
   AlertDialogFooter,
-  Flex,
-  Divider,
   Text,
+  useDisclosure,
 } from "@chakra-ui/react";
-import { SwapState } from "../../Swap";
-import { CloseIcon, WarningIcon } from "@chakra-ui/icons";
-import { LoadingDialog } from "./LoadingDialog";
+import { CloseIcon } from "@chakra-ui/icons";
 
-export enum SwapConfirmationStep {
-  CONFIRM, // Initial step/state
-  SIGN, // When hashpack is awaiting signature
-  TRANSACTION, // When transaction is signed in hashpack and is now executing (NOTE: currently cannot support)
-  ERROR, // When there is an error with the executed transaction (for error dialog)
-}
-interface SwapConfirmationProps {
-  sendSwapTransaction: (payload: any) => void;
-  swapState: SwapState;
-  confirmationStep: SwapConfirmationStep;
-  errorMessage?: string;
-  onErrorMessageDismiss?: () => void | undefined; // fired when error dialog is closed (SwapConfirmationStep.ERROR)
-  onSwapButtonClick?: () => void | undefined; // fired when swap button is clicked (button starts dialog flow)
-  onClose?: () => void | undefined; // fired when swap setting AlertDialog is closed (sendSwapTransaction)
+interface AlertDialogComponentProps {
+  title: string;
+  children: React.ReactNode;
+  openDialogButtonText: string; // trigger to open dialog
+  alertDialogOpen?: boolean; // use this to programtically open/close dialog
+  onAlertDialogOpen?: () => void;
+  onAlertDialogClose?: () => void;
+  modalButtonText?: string; // button at bottom of modal
+  onModalButtonClick?: (event: React.MouseEvent<HTMLButtonElement>) => void;
 }
 
-const SwapConfirmation = (props: SwapConfirmationProps) => {
-  const { sendSwapTransaction, swapState, confirmationStep, errorMessage, onErrorMessageDismiss, onSwapButtonClick } =
-    props;
-  const { isOpen, onOpen, onClose } = useDisclosure();
+/**
+ * <AlertDialogComponent>
+ *  Content to be displayed in dialog body
+ * </AlertDialogComponent>
+ *
+ * Usage note:
+ * This component has built in functionality to open the AlertDialog on click of the trigger and to close
+ * the dialog when the close button in the top right of the modal is clicked.
+ *
+ * If consumers need to programmatically open or close the dialog outside of the aforementioned base use case,
+ * alertDialogOpen prop should be used to set the open/close state. Furthermore, there are callbacks provided
+ * (onAlertDialogOpen, onAlertdialogClose) for when the dialog is opened and closed, which should help with
+ * dynamic state management of opening and closing
+ */
+const AlertDialogComponent = (props: AlertDialogComponentProps) => {
+  const {
+    title,
+    children,
+    openDialogButtonText,
+    alertDialogOpen,
+    onAlertDialogOpen,
+    onAlertDialogClose,
+    modalButtonText,
+    onModalButtonClick,
+  } = props;
   const cancelRef = useRef<any>();
-
-  const [errorDialogOpen, setErrorDialogOpen] = useState(true);
+  const { isOpen, onOpen, onClose } = useDisclosure();
   useEffect(() => {
-    // confirmationStep will be updated by TransactionState slice in HashconnectState
-    // which is updated after transaction is executed
-    if (confirmationStep === SwapConfirmationStep.ERROR) {
-      setErrorDialogOpen(true);
+    // For programatically opening/closing dialog
+    // when alertDialogOpen is set/unset, accordingly call open and close hooks from useDisclosure
+    if (alertDialogOpen) {
+      onOpen();
+      if (onAlertDialogOpen) onAlertDialogOpen();
+    } else {
+      onClose();
+      if (onAlertDialogClose) onAlertDialogClose();
     }
-  }, [confirmationStep]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [alertDialogOpen]);
 
-  // OnClick of Swap button. Opens AlertDialog with swap settings/confirmation and fires
-  // onSwapButtonClick prop
-  const onSwapClick = () => {
+  /** Invoked when dialog is opened. Fires onAlertDialogOpen callback */
+  const openDialog = () => {
     onOpen();
-    if (onSwapButtonClick) onSwapButtonClick();
+    if (onAlertDialogOpen) onAlertDialogOpen();
   };
 
-  const handleSwapTransaction = useCallback(async () => {
-    const { tokenToTrade, tokenToReceive } = swapState;
-    if (tokenToTrade.symbol === undefined || tokenToReceive.symbol === undefined) {
-      console.error("Token types must be selected to Swap tokens.");
-      return;
-    }
-    sendSwapTransaction({
-      tokenToTrade: { ...tokenToTrade },
-      tokenToReceive: { ...tokenToReceive },
-    });
-    // close AlertDialog with swap settings/confirmation and fire onClose prop
+  /** Invoked when dialog is closed. Fires onAlertDialogClose callback */
+  const closeDialog = () => {
     onClose();
-    if (props.onClose) props.onClose();
-  }, [sendSwapTransaction, swapState, onClose, props]);
-
-  /**
-   * Calculates minimum received amount after slippage
-   * Min Received = Receive Amount - (Receive Amount * slippage percentage)
-   */
-  const getMinReceivedAfterSlippage = useCallback(() => {
-    const receiveAmount = swapState.tokenToReceive.amount;
-    const tokenToReceive = swapState.tokenToReceive.symbol;
-    const slippage = +swapState.swapSettings.slippage / 100;
-    return `${(receiveAmount - receiveAmount * slippage).toFixed(7)} ${tokenToReceive}`;
-  }, [swapState]);
-
-  const swapData = useCallback((): { [key: string]: any } => {
-    const { tokenToTrade, tokenToReceive, swapSettings, spotPrice } = swapState;
-    const minReceivedKey = `Minimum received after slippage (${+swapSettings.slippage}%)`;
-    return {
-      pairDetails: {
-        Trade: `${tokenToTrade.amount} ${tokenToTrade.symbol}`,
-        Receive: `${tokenToReceive.amount.toFixed(5)} ${tokenToReceive.symbol}`,
-      },
-      swapSettings: {
-        "Exchange Rate": `1 ${tokenToTrade.symbol} = ${spotPrice?.toFixed(5)} ${tokenToReceive.symbol}`,
-        "Transaction Fee": "0.01 HBAR", // TODO: update this with actual transaction fee
-        "Gas Fee": "0.01 HBAR", // TODO: update this with actual gas fee
-        [minReceivedKey]: getMinReceivedAfterSlippage(),
-      },
-    };
-  }, [swapState, getMinReceivedAfterSlippage]);
+    if (onAlertDialogClose) onAlertDialogClose();
+  };
 
   return (
     <>
-      <Button onClick={onSwapClick} marginTop="0.5rem">
-        Swap
+      <Button onClick={openDialog} marginTop="0.5rem">
+        {openDialogButtonText}
       </Button>
-      {/* // Confirm Swap Dialog */}
       <AlertDialog
         motionPreset="slideInBottom"
         leastDestructiveRef={cancelRef}
@@ -114,79 +91,24 @@ const SwapConfirmation = (props: SwapConfirmationProps) => {
         <AlertDialogContent width={"370px"}>
           <AlertDialogHeader padding={"16px"} display={"flex"} justifyContent={"space-between"} alignItems={"center"}>
             <Text fontSize={"24px"} lineHeight={"29px"} fontWeight={"400"}>
-              Confirm Swap
+              {title}
             </Text>
-            <CloseIcon w={4} h={4} onClick={onClose} cursor={"pointer"} />
+            <CloseIcon w={4} h={4} onClick={closeDialog} cursor={"pointer"} />
           </AlertDialogHeader>
-          <AlertDialogBody padding={"0 16px"}>
-            <dl>
-              <Flex flexDirection={"column"} width={"100%"}>
-                {Object.keys(swapData().pairDetails).map((key: string, i: number) => {
-                  return (
-                    <Flex
-                      justifyContent={"space-between"}
-                      width={"100%"}
-                      marginBottom={i === 0 ? "4px" : "8px"}
-                      fontSize={"18px"}
-                      lineHeight={"22px"}
-                      key={i}
-                    >
-                      <dt style={{ width: "50%" }}>{key}</dt>
-                      <dd style={{ width: "50%", textAlign: "right" }}>{swapData().pairDetails[key]}</dd>
-                    </Flex>
-                  );
-                })}
-                <Divider borderColor={"#000000"} marginBottom={"8px"} />
-                {Object.keys(swapData().swapSettings).map((key: string, i: number) => {
-                  return (
-                    <Flex
-                      justifyContent={"space-between"}
-                      width={"100%"}
-                      marginBottom={"8px"}
-                      fontSize={"14px"}
-                      lineHeight={"17px"}
-                      key={i}
-                    >
-                      <dt style={{ width: "50%" }}>{key}</dt>
-                      <dd style={{ width: "50%", textAlign: "right" }}>{swapData().swapSettings[key]}</dd>
-                    </Flex>
-                  );
-                })}
-              </Flex>
-            </dl>
-          </AlertDialogBody>
+          <AlertDialogBody padding={"0 16px"}>{children}</AlertDialogBody>
           <AlertDialogFooter padding={"16px"}>
-            <Button width={"100%"} ref={cancelRef} onClick={handleSwapTransaction}>
-              Confirm Swap
-            </Button>
+            {modalButtonText ? (
+              <Button width={"100%"} ref={cancelRef} onClick={onModalButtonClick}>
+                {modalButtonText}
+              </Button>
+            ) : (
+              ""
+            )}
           </AlertDialogFooter>
         </AlertDialogContent>
       </AlertDialog>
-      {confirmationStep === SwapConfirmationStep.SIGN ? (
-        // Sign transaction dialog
-        <LoadingDialog isOpen={true} message={"Please confirm the swap in your wallet to proceed"} />
-      ) : confirmationStep === SwapConfirmationStep.ERROR ? (
-        <LoadingDialog
-          isOpen={errorDialogOpen}
-          message={errorMessage ?? ""}
-          icon={<WarningIcon h={10} w={10} />}
-          buttonConfig={{
-            text: "Dismiss",
-            onClick: () => {
-              setErrorDialogOpen(false);
-              if (onErrorMessageDismiss) onErrorMessageDismiss();
-            },
-          }}
-        />
-      ) : confirmationStep === SwapConfirmationStep.TRANSACTION ? (
-        // Transaction processing dialog
-        // TODO: add transaction id
-        <LoadingDialog isOpen={true} message={"Processing Transaction ABC12345. This may take several minutes."} />
-      ) : (
-        ""
-      )}
     </>
   );
 };
 
-export { SwapConfirmation };
+export { AlertDialogComponent };
